requirements.txt
fastapi==0.103.1
imgaug==0.4.0
matplotlib==3.7.1
numpy==1.23.5
opencv_python==4.5.5.64
Pillow==9.4.0
Pillow==10.0.1
progress==1.6
ptflops==0.7
Requests==2.31.0
scipy==1.11.2
torch==1.13.1
torchsummary==1.5.1
torchvision==0.14.1
uvicorn==0.11.8